Refactor system locks and add tests
